Remove `ENTITY_TYPE` args, create `adapters` folder under `bug_gen`
