+++ swesmith/bug_gen/adapters/__init__.py
@@ -1,0 +1,7 @@
+from swesmith.bug_gen.adapters.python import (
+    py_get_entities_from_file,
+)
+
+get_entities_from_file = {
+    "py": py_get_entities_from_file,
+}
