+++ swesmith/bug_gen/utils.py
@@ -1,72 +1,17 @@
-import ast
 import os
 import subprocess
 
-from dataclasses import dataclass
-from typing import Any
 from dotenv import load_dotenv
 from itertools import combinations
+from pathlib import Path
+from swesmith.bug_gen.adapters import get_entities_from_file
 from swesmith.constants import TEMP_PATCH
-from swesmith.utils import generate_hash
+from swesmith.utils import BugRewrite, CodeEntity
 
 load_dotenv()
 
 
 DEVNULL = {"stdout": subprocess.DEVNULL, "stderr": subprocess.DEVNULL}
-ENTITY_TYPES = {
-    "class": [ast.ClassDef],
-    "func": [ast.FunctionDef],
-    "object": [ast.ClassDef, ast.FunctionDef],
-}
-
-
-class BugRewrite:
-    cost: float = 0
-    explanation: str = ""
-    output: str
-    rewrite: str
-    strategy: str
-
-    def __init__(
-        self,
-        rewrite: str,
-        explanation: str,
-        strategy: str,
-        cost: float = 0,
-        output: str = "",
-    ):
-        self.rewrite = rewrite
-        self.explanation = explanation
-        self.cost = cost
-        self.strategy = strategy
-        self.output = output
-
-    def get_hash(self) -> str:
-        """Generates a hash for the bug rewrite."""
-        return generate_hash(self.rewrite)
-
-    def to_dict(self) -> dict[str, Any]:
-        """Converts the bug rewrite to a dictionary."""
-        return {
-            "cost": self.cost,
-            "explanation": self.explanation,
-            "output": self.output,
-            "rewrite": self.rewrite,
-            "strategy": self.strategy,
-        }
-
-
-@dataclass
-class CodeEntity:
-    """Data class to hold information about a code entity (e.g. function, class)."""
-
-    file_path: str
-    indent_level: int
-    indent_size: int
-    line_end: int
-    line_start: int
-    src_code: Any
-    src_node: Any
 
 
 def apply_code_change(candidate: CodeEntity, bug: BugRewrite) -> None:
@@ -122,7 +67,6 @@ def apply_patches(repo: str, patch_files: list[str]) -> str | None:
 
 def extract_entities_from_directory(
     directory_path: str,
-    entity_type: str,
     exclude_tests: bool = True,
     max_entities: int = -1,
 ) -> list[CodeEntity]:
@@ -130,7 +74,6 @@ def extract_entities_from_directory(
     Extracts entities (functions, classes, etc.) from Python files in a directory.
     Args:
         directory_path (str): Path to the directory to scan.
-        entity_type (str): Type of entity to extract.
         exclude_tests (bool): Whether to exclude test files and directories.
     Returns:
         List[CodeEntity]: List of CodeEntity objects containing entity information.
@@ -138,8 +81,6 @@ def extract_entities_from_directory(
     entities = []
     for root, _, files in os.walk(directory_path):
         for file in files:
-            if not file.endswith(".py"):
-                continue
             if exclude_tests and any(
                 [x in root for x in ["/tests", "/test", "/testing"]]
             ):
@@ -155,17 +96,13 @@ def extract_entities_from_directory(
             except:
                 continue
 
-            try:
-                tree = ast.parse(file_content, filename=file_path)
-            except SyntaxError:
+            file_ext = Path(file_path).suffix[1:]
+            if file_ext not in get_entities_from_file:
                 continue
+            get_entities_from_file[file_ext](
+                entities, file_content, file_path, max_entities
+            )
 
-            for node in ast.walk(tree):
-                if not any([isinstance(node, x) for x in ENTITY_TYPES[entity_type]]):
-                    continue
-                entities.append(get_entity_from_node(node, file_content, file_path))
-                if max_entities != -1 and len(entities) >= max_entities:
-                    return entities
     return entities
 
 
@@ -177,61 +114,6 @@ def get_combos(items, r, max_combos) -> list[tuple]:
         if max_combos != -1 and len(all_combos) >= max_combos:
             break
     return sorted(all_combos, key=len)
-
-
-def get_entity_from_node(
-    node: ast.AST, file_content: str, file_path: str
-) -> CodeEntity:
-    """Turns an AST node into a CodeEntity object."""
-    start_line = node.lineno  # type: ignore[attr-defined]
-    end_line = (
-        node.end_lineno if hasattr(node, "end_lineno") else None  # type: ignore[attr-defined]
-    )
-
-    if end_line is None:
-        # Calculate end line manually if not available (older Python versions)
-        end_line = (
-            start_line
-            + len(
-                ast.get_source_segment(file_content, node).splitlines()  # type: ignore[attr-defined]
-            )
-            - 1
-        )
-
-    source_code = ast.get_source_segment(file_content, node)
-
-    # Get the line content for the source definition
-    source_line = file_content.splitlines()[start_line - 1]
-    leading_whitespace = len(source_line) - len(source_line.lstrip())
-
-    # Determine the number of spaces per tab
-    indent_size = 4  # Default fallback
-    if "\t" in file_content:
-        indent_size = source_line.expandtabs().index(source_line.lstrip())
-
-    # Calculate indentation level
-    indentation_level = (
-        leading_whitespace // indent_size if leading_whitespace > 0 else 0
-    )
-
-    # Remove indentation from source source code
-    assert source_code is not None
-    lines = source_code.splitlines()
-    dedented_source_code = [lines[0]]
-    for line in lines[1:]:
-        # Strip leading spaces equal to indentation_level * indent_size
-        dedented_source_code.append(line[indentation_level * indent_size :])
-    source_code = "\n".join(dedented_source_code)
-
-    return CodeEntity(
-        file_path=file_path,
-        indent_level=indentation_level,
-        indent_size=indent_size,
-        line_end=end_line,
-        line_start=start_line,
-        src_code=source_code,
-        src_node=node,
-    )
 
 
 def get_patch(repo: str, reset_changes: bool = False):
